requirements/development.txt
-r base.txt
flake8==3.7.8
pytest==5.1.0
pytest-django==3.5.1
